Update: 로그인 안했을때 로컬 스토리지 구현 - 로그인 했을때 로그인한 userId로 바꿔서 insert 시켜줘야함

+++ src/hooks/mutation/useCartsMutation/useCartsMutation.ts
@@ -1,11 +1,29 @@
 import { deleteAllCartByUser, deleteCartByUser, patchCartByUser, postCartByUser } from '@/api/cart';
+import { getProductById } from '@/api/product';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 
 const useCartsMutation = () => {
   const queryClient = useQueryClient();
+  const localCarts = JSON.parse(localStorage.getItem('carts') || '[]');
   const addMutation = useMutation({
-    mutationFn: async ({ productId, userId }: { productId: number; userId: string }) => {
-      await postCartByUser(productId, userId);
+    mutationFn: async ({ productId, userId }: { productId: number; userId: string | null }) => {
+      if (userId) {
+        await postCartByUser(productId, userId);
+      } else {
+        const product = await getProductById({ params: { productId } });
+        localStorage.setItem(
+          'carts',
+          JSON.stringify(
+            localCarts.concat({
+              productId,
+              userId,
+              count: 1,
+              select: false,
+              Products: product
+            })
+          )
+        );
+      }
     },
     onSuccess: (data, variable) => {
       queryClient.refetchQueries({ queryKey: ['Carts', variable.userId] });
@@ -20,12 +38,23 @@ const useCartsMutation = () => {
       cal
     }: {
       productId: number;
-      userId: string;
+      userId: string | null;
       count: number;
       cal: boolean;
     }) => {
-      if (cal) await patchCartByUser(productId, userId, count + 1);
-      else await patchCartByUser(productId, userId, count - 1);
+      if (userId) {
+        if (cal) await patchCartByUser(productId, userId, count + 1);
+        else await patchCartByUser(productId, userId, count - 1);
+      } else {
+        const updatedCarts = localCarts.map((cart: Cart) => {
+          if (cart.productId === productId) {
+            return { ...cart, count: cal ? count + 1 : count - 1 };
+          }
+          return cart;
+        });
+
+        localStorage.setItem('carts', JSON.stringify(updatedCarts));
+      }
     },
     onMutate: async (variable) => {
       await queryClient.cancelQueries({ queryKey: ['Carts', variable.userId] });
@@ -47,16 +76,19 @@ const useCartsMutation = () => {
     }
   });
   const deleteAllMutation = useMutation({
-    mutationFn: async ({ userId }: { userId: string }) => {
-      await deleteAllCartByUser(userId);
+    mutationFn: async ({ userId }: { userId: string | null }) => {
+      if (userId) await deleteAllCartByUser(userId);
+      else localStorage.removeItem('carts');
     },
     onSuccess: (data, variable) => {
       queryClient.invalidateQueries({ queryKey: ['Carts', variable.userId] });
     }
   });
   const deleteMutation = useMutation({
-    mutationFn: async ({ productId, userId }: { productId: number; userId: string }) => {
-      await deleteCartByUser(productId, userId);
+    mutationFn: async ({ productId, userId }: { productId: number; userId: string | null }) => {
+      if (userId) await deleteCartByUser(productId, userId);
+      else
+        localStorage.setItem('carts', JSON.stringify(localCarts.filter((cart: Cart) => cart.productId !== productId)));
     },
     onSuccess: (data, variable) => {
       queryClient.invalidateQueries({ queryKey: ['Carts', variable.userId] });
